src/test/kotlin/dev/shtanko/algorithms/leetcode/LargestUniqueNumberTest.kt
package dev.shtanko.algorithms.leetcode

import java.util.stream.Stream
import org.junit.jupiter.api.Assertions.assertEquals
import org.junit.jupiter.api.extension.ExtensionContext
import org.junit.jupiter.params.ParameterizedTest
import org.junit.jupiter.params.provider.Arguments
import org.junit.jupiter.params.provider.ArgumentsProvider
import org.junit.jupiter.params.provider.ArgumentsSource

internal abstract class LargestUniqueNumberTest(val strategy: LargestUniqueNumber) {

    internal class InputArgumentsProvider : ArgumentsProvider {
        override fun provideArguments(context: ExtensionContext?): Stream<out Arguments> = Stream.of(
            Arguments.of(
                intArrayOf(5, 7, 3, 9, 4, 9, 8, 3, 1),
                8
            )
        )
    }

    @ParameterizedTest
    @ArgumentsSource(InputArgumentsProvider::class)
    internal fun `largest unique number test`(arr: IntArray, expected: Int) {
        val actual = strategy.perform(arr)
        assertEquals(expected, actual)
    }
}

internal class LargestUniqueNumberBruteForceTest : LargestUniqueNumberTest(LargestUniqueNumberBruteForce())
internal class LargestUniqueNumberHashMapTest : LargestUniqueNumberTest(LargestUniqueNumberHashMap())
